+++ ugs-pendant/src/main/webapp/src/components/FeedRate.tsx
@@ -18,8 +18,8 @@ const FeedRate = ({ value, onChange }: Props) => {
       className="stepSize"
       title={value}
     >
-      {options.map((option, index) => (
-        <Dropdown.Item key={index} as="button" onClick={() => onChange(option)}>
+      {options.map(option => (
+        <Dropdown.Item key={option} as="button" onClick={() => onChange(option)}>
           {option} {units.toLocaleLowerCase() + "/min"}
         </Dropdown.Item>
       ))}
